requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.115.2
fastapi-login
h11==0.14.0
httpx==0.26
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
pendulum==2.1.2
pydantic==2.9.2
pydantic_core==2.23.4
pydantic_settings==2.6.0
pymongo==4.4.1
python-dateutil==2.8.2
python-multipart==0.0.5
pytzdata==2020.1
requests==2.27.1
six==1.16.0
sniffio==1.3.0
starlette==0.40.0
typing_extensions==4.12.2
urllib3==1.26.16
uvicorn==0.20.0
pyopenssl 
requests[security]
ndg-httpsclient 
pyasn1
supabase
psycopg2
python-dotenv
python-jose
ruff
pre-commit
